# app/mgmt/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "此文档由gin-swagger自动生成",
        "title": "xxim-server HTTP API 文档",
        "contact": {
            "name": "showurl"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:6799",
    "basePath": "/api",
    "paths": {
...
        "/ms/add/admin": {
            "post": {
                "description": "使用此接口添加管理员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员相关接口"
                ],
                "summary": "添加管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.AddMSUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.AddMSUserResp"
                        }
                    }
                }
            }
        },
...
        "/ms/delete/admin": {
            "post": {
                "description": "使用此接口删除管理员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员相关接口"
                ],
                "summary": "删除管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.DeleteMSUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.DeleteMSUserResp"
                        }
                    }
                }
            }
        },
...
        "/ms/get/admin/detail": {
            "post": {
                "description": "使用此接口获取管理员详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员相关接口"
                ],
                "summary": "获取管理员详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetMSUserDetailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/mshandler.getAllMSUserListRespUser"
                        }
                    }
                }
            }
        },
        "/ms/get/admin/detail/self": {
            "post": {
                "description": "使用此接口获取自己的管理员详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员相关接口"
                ],
                "summary": "获取自己的管理员详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.GetSelfMSUserDetailResp"
                        }
                    }
                }
            }
        },
        "/ms/get/admin/list/all": {
            "post": {
                "description": "使用此接口获取所有管理员列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员相关接口"
                ],
                "summary": "获取所有管理员列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetAllMSUserListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/mshandler.getAllMSUserListResp"
                        }
                    }
                }
            }
        },
...
        "/ms/switch/admin/status": {
            "post": {
                "description": "使用此接口切换管理员状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员相关接口"
                ],
                "summary": "切换管理员状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.SwitchMSUserStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.SwitchMSUserStatusResp"
                        }
                    }
                }
            }
        },
...
        "/ms/update/admin": {
            "post": {
                "description": "使用此接口更新管理员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员相关接口"
                ],
                "summary": "更新管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateMSUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateMSUserResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mshandler.getAllMSUserListResp": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "pageNo": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mshandler.getAllMSUserListRespUser"
                    }
                }
            }
        },
        "mshandler.getAllMSUserListRespUser": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "string"
                },
                "isDisable": {
                    "description": "是否禁用: [0=否, 1=是]",
                    "type": "boolean"
                },
                "isMultipoint": {
                    "description": "多端登录: [0=否, 1=是]",
                    "type": "boolean"
                },
                "lastLoginIp": {
                    "description": "最后登录IP",
                    "type": "string"
                },
                "lastLoginTime": {
                    "description": "最后登录时间",
                    "type": "string"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "role": {
                    "description": "角色",
                    "type": "string"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "username": {
                    "description": "账号",
                    "type": "string"
                }
            }
        },
...
        "pb.AddMSUserReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "user": {
                    "$ref": "#/definitions/pb.MSUser"
                }
            }
        },
        "pb.AddMSUserResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.CommonReq": {
            "type": "object",
            "properties": {
                "appVersion": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "deviceId": {
                    "type": "string"
                },
                "deviceModel": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "osVersion": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "userAgent": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "pb.CommonResp": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/pb.CommonResp_Code"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "pb.CommonResp_Code": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                7,
                8,
                9,
                10
            ],
            "x-enum-comments": {
                "CommonResp_AlertError": "alert 错误 只有一个确定按钮",
                "CommonResp_AuthError": "鉴权错误 // 应该退出登录",
                "CommonResp_ForbiddenError": "无权限",
                "CommonResp_InternalError": "内部错误",
                "CommonResp_NeedSecondPasswordError": "需要输入二级密码/设置二级密码",
                "CommonResp_RequestError": "请求错误",
                "CommonResp_RetryError": "alert 错误 有一个取消按钮 和一个重试按钮",
                "CommonResp_ToastError": "toast 错误 只有 message",
                "CommonResp_UnknownError": "未知 error"
            },
            "x-enum-varnames": [
                "CommonResp_Success",
                "CommonResp_UnknownError",
                "CommonResp_InternalError",
                "CommonResp_RequestError",
                "CommonResp_AuthError",
                "CommonResp_ToastError",
                "CommonResp_AlertError",
                "CommonResp_RetryError",
                "CommonResp_ForbiddenError",
                "CommonResp_NeedSecondPasswordError"
            ]
        },
...
        "pb.DeleteMSUserReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "pb.DeleteMSUserResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.GetAllMSUserListReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "page": {
                    "$ref": "#/definitions/pb.Page"
                }
            }
        },
...
        "pb.GetMSUserDetailReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "pb.GetSelfMSUserDetailResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                },
                "permissions": {
                    "description": "权限集合: [[*]=\u003e所有权限, ['article:add']=\u003e部分权限]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user": {
                    "$ref": "#/definitions/pb.MSUser"
                }
            }
        },
...
        "pb.MSUser": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "createdAtStr": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isDisable": {
                    "type": "boolean"
                },
                "lastLoginIp": {
                    "type": "string"
                },
                "lastLoginTime": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "integer"
                },
                "updatedAtStr": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "pb.Page": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "pb.SwitchMSUserStatusReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "pb.SwitchMSUserStatusResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.UpdateMSUserReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "user": {
                    "$ref": "#/definitions/pb.MSUser"
                }
            }
        },
        "pb.UpdateMSUserResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
    }
}